refactor: Extract app elements
